wedge in the cg3 framebuffer instead of tcx
This is a hack to get a cg3 framebuffer for sparc.  This won't
work with openbios, so you need the proprietary system firmware
rom image and the rom image for the cg3 card.  The cg3 rom is
expected to be named cg3.bin and be in your current working
directory.